Created an abstract hosted connection service that has the autohost, start/stopHostingOnConnection support built into it.  This is a very common things for connection based services and I got tired of cutting/pasting it all the time. RpcHostedService was modified to extend this base class instead of the more basic one.
